src/requirements.txt
streamlit
google-generativeai
python-dotenv
pymupdf
scikit-learn
numpy